requirements.txt
flask == 1.1.2
pillow == 7.1.2
Flask-WTF == 0.14.3
Flask-Uploads == 0.2.1
Werkzeug == 0.16.1